Fix sender table having a lot of duplicated senders (bug 1245)
Commit 20f446a4 introduced a bug, where the unique index on sender
(sender, realname, avatarurl) did not catch inserts where realname or
avatarurl were NULL, because the columns lacked the NOT NULL constraint.
Unique indexes do not consider NULL equal to NULL.

This means a lot of new senderid's were created (one for each insert in
the sender table) where one of these values was null, which bloated the DB
unnecessarily, and may have caused other bad behavior.

This commit:
- adds NOT NULL constraints on the sender table realname and avatarurl columns
- changes the sender insert statements to coalesce the value with an
  empty string (otherwise there are errors where it attempts to insert
  NULL)
- adds upgrade scripts

The upgrade scripts are quite complex:
- first they drop the unique index on sender
- update all null values on modified columns to empty string
- update the sender table schema to add NOT NULL constraints
- create a temp table that holds all rows that would violate the unique
  index, with good and bad senderid, that are used later
- triggers on backlog and sender tables are disabled or deleted,
  to speed up the operation
- backlog table bad senderid's are updated to good senderid's
- rows from sender with bad senderid's are deleted
- triggers are reenabled or recreated
- tables sender and backlog are reindexed
- unique index on sender is recreated

I opted to not put the statements into separate numbered files in the
upgrade directory, as they would get put into separate transactions,
which could leave the db in a hard to recover state if any of them are
unsuccessful. Better to leave the db untouched in the previous version.